Simplify the instructions to get the experimental branch
When you `dev` a package, if it isn't updated manually through git it will stay on the commit you dev'ed it at.  However, when adding a branch, Pkg automatically searches for the latest commit on that branch.  This simplifies the workflow when you only want a branch, without having to download the entire git repo.